This PR improves the type safety of the Java FFI and fixes a bug with pointers to Nit objects at the same time.

* The main change is the introduction of the Java class `NitObject` used as a general type for Nit objects referenced from Java code. It replaces the use of a simple `int` holding the pointer value, which was the source of the Java FFI bug.

* Change internal pointers to Nit object to use the Java primitive types `long` instead of `int`, addressing to the same pointer bug. Only low-level services should still use `long` this way, in this case it is the support for Android that relies mostly on the FFI with C.

* Change the annotation `extra_java_files` to accept the full name of Java classes (package + class) instead of the path to the Java source file. The compile still looks for the file in the same directory as the Nit module.

* Clear up the documentation of the class `ExternFile` for `filename` to be relative to the compilation folder. This required updating previous usages.

Pull-Request: #2740

+++ lib/android/service/NitService.java
@@ -22,7 +22,7 @@ import android.os.IBinder;
 // Service implemented in Nit
 public class NitService extends Service {
 
-	protected int nitService = 0;
+	protected long nitService = 0;
 
 	static {
 		System.loadLibrary("nit_app");
@@ -51,8 +51,8 @@ public class NitService extends Service {
 		return null;
 	}
 
-	protected native int nitNewService();
-	protected native int nitOnStartCommand(int nitService, Intent intent, int flags, int id);
-	protected native void nitOnCreate(int nitService);
-	protected native void nitOnDestroy(int nitService);
+	protected native long nitNewService();
+	protected native int nitOnStartCommand(long nitService, Intent intent, int flags, int id);
+	protected native void nitOnCreate(long nitService);
+	protected native void nitOnDestroy(long nitService);
 }
